# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FaceGuard - Face Recognition Access & Attendance System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /zones:
    get:
      operationId: listZones
      tags: [zones]
      summary: List all zones/rooms
      responses:
        "200":
          description: List of zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Zone"
    post:
      operationId: createZone
      tags: [zones]
      summary: Create a zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoneInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
  /zones/{id}:
    patch:
      operationId: updateZone
      tags: [zones]
      summary: Update zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoneUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
    delete:
      operationId: deleteZone
      tags: [zones]
      summary: Delete zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ACCESS RULES
...
components:
  schemas:
...
    Zone:
      type: object
      required: [id, name, accessLevel, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        accessLevel:
          type: string
          enum: [public, restricted, secure]
        cameraCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    ZoneInput:
      type: object
      required: [name, accessLevel]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        accessLevel:
          type: string
          enum: [public, restricted, secure]
    ZoneUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        accessLevel:
          type: string
          enum: [public, restricted, secure]
...