# lib/api-zod/src/generated/types/zoneUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FaceGuard - Face Recognition Access & Attendance System
 * OpenAPI spec version: 0.1.0
 */
import type { ZoneUpdateAccessLevel } from './zoneUpdateAccessLevel';

export interface ZoneUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  accessLevel?: ZoneUpdateAccessLevel;
}